refactor: rename sidebar component
