+++ src/shared/components/user-area-sidebar.tsx
